[9] Add missing after resolved conflict
